# docs/swagger.yaml
definitions:
...
  controllers.HistoryRatingInput:
    properties:
      game_id:
        type: integer
      rating_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  controllers.RatingInput:
    properties:
      value:
        type: string
    type: object
...
  models.Game:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  models.HistoryRating:
    properties:
      created_at:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      rating_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Rating:
    properties:
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      value:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /HistoryRating/{id}:
    delete:
      description: Delete a HistoryRating by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: HistoryRating id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one HistoryRating.
      tags:
      - HistoryRating
  /HistoryRatings:
    get:
      description: Get a list of HistoryRatings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.HistoryRating'
            type: array
      summary: Get all HistoryRatings.
      tags:
      - HistoryRating
    post:
      description: Creating a new HistoryRating.
      parameters:
      - description: the body to create a new HistoryRating
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.HistoryRatingInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoryRating'
      security:
      - BearerToken: []
      summary: Create New HistoryRating.
      tags:
      - HistoryRating
  /HistoryRatings/{id}:
    get:
      description: Get a HistoryRating by id.
      parameters:
      - description: HistoryRating id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoryRating'
      summary: Get HistoryRating.
      tags:
      - HistoryRating
    patch:
      description: Update HistoryRating by id.
      parameters:
      - description: HistoryRating id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an HistoryRating
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.HistoryRatingInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HistoryRating'
      security:
      - BearerToken: []
      summary: Update HistoryRating.
      tags:
      - HistoryRating
  /HistoryRatings/{id}/Games:
    get:
      description: Get HistoryRating by Game.
      parameters:
      - description: Game id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
      summary: Get HistoryRating By Game.
      tags:
      - HistoryRating
  /Rating:
    get:
      description: Get a list of Rating.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Rating'
            type: array
      summary: Get all Rating.
      tags:
      - Rating
    post:
      description: Creating a new Rating.
      parameters:
      - description: the body to create a new Rating
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RatingInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rating'
      security:
      - BearerToken: []
      summary: Create New Rating.
      tags:
      - Rating
  /Rating/{id}:
    delete:
      description: Delete a Rating by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Rating id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Rating.
      tags:
      - Rating
    patch:
      description: Update Rating by id.
      parameters:
      - description: Rating id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age Rating Rating
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RatingInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Rating'
      security:
      - BearerToken: []
      summary: Update Rating.
      tags:
      - Rating
...
swagger: "2.0"